build: Add Muen System
This commit adds the infrastructure needed to build and run an emulated
Muen System with S-Keyfender and other required components.

Note that the built system is not yet functional, further work to
complete this will come in later changes.

The src/muen submodule is updated to
598ade47383216aea37054e9992f6005f3a96b4b.

A src/zlib submodule is added containing ZLib, which is required to
cross-build Git for the Muen initramfs.