Update NitroHSM Muen system to latest upstream devel branch
The src/muen submodule is updated to point directly the upstream "devel"
branch as of 26f5f504bfe40203ca9a7a4a948feb54bfe3e93a. Our "local"
"devel-nitrohsm" branch is no longer used.

In order to be compatible with the upstream changes, and to allow for
easier tracking of the upstream "devel" branch in the future, the
following changes are made to the NitroHSM/Muen build system:

- Muen commit 0ff0abb57a656c28cabb9b43215c4e1740ae7232 introduced an
improved build system for Linux components. Muen-specific kernel modules
are now built as Git submodules and a separate initramfs image with
kernel modules matching the Linux kernel is produced. This allows us to
override the upstream Muen use of Buildroot by just passing a suitable
INITRAMFS= to the Muen build, and *both* initramfs images will be
combined into a single image visible to the Linux userland at boot time.
- U-Root is added as a submodule to this repository, along with the
required build recipes to produce the image passed as INITRAMFS= to the
Muen build system.
- Muen system policy sources for NitroHSM are added directly to this
repository in src/muen-policy and are passed to the Muen build system.

In summary, this is effectively a "V2" of the NitroHSM/Muen build, and
puts us in a much better position going forward as we no longer need to
carry any local changes to the src/muen submodule.