Fix shallow cloning of submodules
The original repository of the git project doesn't allow direct
fetches of tagged commits.  However the mirror on Github does, which
this commit is using instead.  Also a "--depth 1" option is added to
the submodule update command, which - although .gitmodule sets
shallow=true - unexpectedy reduces the amount of fetched objects
significantly.

Closes: #15
Signed-off-by: Sven Anderson <sven@anderson.de>