Muen: Use Coreboot for emulated system
This switches the emulated Muen system (MODE=muen) to use Coreboot 4.12
with GRUB 2 as a payload instead of the existing "GRUB 2 as kernel"
setup. This is closer to the arrangement that will be used on the target
hardware.

Note that the Coreboot ROM is intentionally NOT built from source,
instead binaries with instructions on how to reproduce them are provided
in coreboot/.