Muen: Update system policy to reflect upstream changes
This updates the NitroHSM system policy to reflect upstream changes,
specifically those related to the new Integrated Subject Initialisation
and Reset [1].

In order to simplify the changes required, we remove subject_nic_sl
entirely, replacing it with a setup identical to that of storage_linux.

All our subjects are put under self-control, and no controller subject
is used.