Muen: Populate /data on system boot from template
This change is in preparation for being able to run the Git server as a
non-root user. Instead of populating the /data partition at "install"
time, or during a RESET event, we produce a template CPIO archive which
is embedded into the U-root initramfs, and is extracted on first system
boot with an empty /data partition.

While this is a somewhat round-about way of doing it, it is actually the
least amount of work required to be able to explicitly specify ownership
of files created in /data in an explicit and declarative manner.

Additionally, the preparation of the template for populating /data is
now done only by the Makefile, whereas previously it was duplicated in
the installer, uinit RESET flow and Makefile.