keyfender: Fix build breakage (webmachine pin)
webmachine was pinned to d1e0909769307870525ab5d9cd83cc29d8745df7 from
https://github.com/roburio/ocaml-webmachine.git. That commit was part of
the 'post-create' branch, which appears to have been rebased.

I have re-pushed the original branch up to that commit from a local
copy to a new 'nitrokey' branch in the same repository, and updated the
keyfender OPAM file to match.