CI, build: Add Docker container, build using it
This adds a Dockerfile describing a Docker image suitable for building
NitroHSM to tools/docker. The image is based on
ocaml/opam:debian-10-ocaml-4.11-flambda with additions required to build
a NitroHSM Muen system. See comments in
tools/docker/Dockerfile.nethsm-builder for details.

This Docker image is now pushed to Docker Hub as mato/nethsm-builder, so
switch the Gitlab CI to use it instead of the stock ocaml/opam image.