Muen system: Use TPM for "Device ID"
Store a 256-bit "Device ID" in the TPM NVRAM. If the TPM has not been
provisioned with one, a random "Device ID" is generated and stored on
first boot when the "Device ID" is first requested from S-Platform.

Note that after this change the Muen system will no longer boot without
a TPM present.

THIS IMPLEMENTATION IS INTENTIONALLY INSECURE and does not protect the
TPM NV Index of the "Device ID" in any way.

When a "RESET" request is sent to S-Platform from S-Keyfender, the
"Device ID" will be de-provisioned (removed from the TPM) before
rebooting the system. This functionality may not be desirable or
possible with a production TPM-based design, but is currently present to
ease testing and to be able to return a unit to a fully "unprovisioned"
state without requiring extra tricks/tools.