Store: after every 100 pushes (remove/set), do a git clone
This removes all references to the old repository state, and since the clone is
done with a depth of 1, it should reduce the memory footprint of S-keyfender.

This addresses #100 at the price that mutating the store gets more expensive
(time and local network wise), which should be fine (since most operations will
be sign and decrypt).