Re-enable downgrade prevention test, fixes #132
This was disabled in !138 / 9b1be8d63000a4b6bb0b1e1c7eb3e047d6f9bfad.

Since the version is now derived from CHANGES.md (since
163d1c1cccb89b91582dd2901db7fc900e01829a), which contains 0.9 at the moment,
the test uses "-1.5" as version number. This is considered a major version
downgrade, which is prevented as discussed in the system design.