keyfender: bind the unlock passphrase to the device-key
An Unlock Key encrypts the Domain Key, that is stored in one of the
two slots in the Device Key Store.  In Unattended Mode the Unlock Key
is identical to the Device Key.  In Attended Mode, the Unlock
Passphrase is hashed with SCrypt and a stored salt.  The Unlock Key
then is SHA256( Passphrase Hash || Device Key) (|| = concatenation).

Signed-off-by: Sven Anderson <sven@anderson.de>