Add API spec validation using openapi-generator-cli
In !335 we realized that the existing OpenAPI validation CI job does not
detect duplicate keys.  Apparently, this is a known limitation of the
Swagger validator:
        https://github.com/swagger-api/validator-badge/issues/60

To catch cases like this one and potentially even others, this patch
adds another CI job that uses the openapi-generator-cli, the parser that
is also used for SDK generation, to validate the API spec.

https://git.nitrokey.com/nitrokey/nethsm/nethsm/-/merge_requests/335