Make namespace a required argument for key operations
Almost all key operations already pass the namespace argument.  But as
it is optional, it can easily be missed in cases where it should be set,
see for example https://github.com/Nitrokey/nethsm/issues/13.

This patch makes the namespace argument for all key operations required.
This leads to somewhat more verbose code when explicitly using keys
without a namespace, mostly in the tests, but also reduces the risk for
bugs caused by an omitted argument.