Implement batching for etcd store with transactions
KV.batch is now implemented for Etcd_store by batching all write/deletes
in blocks of 128 (max) requests sent as etcd transactions, which are
atomic. This is *much* faster than doing unary requests one by one,
resulting for example in a dramatic speedup for restoring backups. This
can be sped up even more if we allow ourselves to increase etcd's
max-txn-ops to 256, allowing a restore of 1 million keys in less than 20
seconds (which was the point of this commit).

To comply with the original semantics of KV.batch, which require that
reads during a batch return the value before starting the batch, all reads are
performed first, without batching, while sets/deletes are queued. Then
all prepared transactions are sent in serial, and waited upon to
finalize the batch.

One added constraint from etcd is that two operations in a batch *cannot* target
the same key.